+++ SupportPortalInfrastructure/Repositories/ProjectNoteRepository.cs
@@ -12,7 +12,7 @@ public class ProjectNoteRepository : GenericRepository<ProjectNoteEntity>, IProj
 
     public ProjectNoteRepository(SupportPortalDBContext context) : base(context) { }
 
-    public async Task<IEnumerable<ProjectNoteEntity>> GetByProjectIdAsync(int projectId, CancellationToken cancellationToken = default) =>
+    public async Task<IEnumerable<ProjectNoteEntity>> GetByProjectIdAsync(Int64 projectId, CancellationToken cancellationToken = default) =>
         await _dbSet
             .Where(pn => pn.ProjectId == projectId)
             .ToListAsync(cancellationToken);
